unignore tests
