+++ javalin-openapi/src/test/java/io/javalin/plugin/openapi/TestOpenApiAnnotations_Java.java
@@ -10,6 +10,7 @@ import io.javalin.plugin.openapi.annotations.OpenApi;
 import io.javalin.plugin.openapi.annotations.OpenApiContent;
 import io.javalin.plugin.openapi.annotations.OpenApiParam;
 import io.javalin.plugin.openapi.annotations.OpenApiResponse;
+import io.javalin.plugin.openapi.utils.VersionIssuesUtil;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import kotlin.Unit;
@@ -177,29 +178,34 @@ class JavaFieldReference {
 class JavaLambdaFieldReference {
     @OpenApi(
         responses = {@OpenApiResponse(status = "200")})
-    public final Handler handler = ctx -> { };
+    public final Handler handler = ctx -> {
+    };
 }
 
 class JavaMultipleFieldReferences {
     @OpenApi(
         path = "/test1", // parameter needed to due ambiguity
         responses = {@OpenApiResponse(status = "200")})
-    public final Handler handler1 = ctx -> { };
+    public final Handler handler1 = ctx -> {
+    };
 
     @OpenApi(
         path = "/test2", // parameter needed to due ambiguity
         responses = {@OpenApiResponse(status = "200")})
-    public final Handler handler2 = ctx -> { };
+    public final Handler handler2 = ctx -> {
+    };
 
     @OpenApi(
         method = HttpMethod.PUT,
         responses = {@OpenApiResponse(status = "200")})
-    public final Handler putHandler = ctx -> { };
+    public final Handler putHandler = ctx -> {
+    };
 
     @OpenApi(
         method = HttpMethod.DELETE,
         responses = {@OpenApiResponse(status = "200")})
-    public final Handler deleteHandler = ctx -> { };
+    public final Handler deleteHandler = ctx -> {
+    };
 }
 
 class JavaStaticLambdaFieldReference {
@@ -207,7 +213,8 @@ class JavaStaticLambdaFieldReference {
         path = "/test",
         method = HttpMethod.GET,
         responses = {@OpenApiResponse(status = "200")})
-    public static final Handler handler = ctx -> { };
+    public static final Handler handler = ctx -> {
+    };
 }
 
 class CustomOuterClassHandler implements Handler {
@@ -225,13 +232,15 @@ class CustomOuterClassHandler implements Handler {
 
 class JavaOuterClassFieldReference {
     @OpenApi(responses = {@OpenApiResponse(status = "200")})
-    public final Handler handler = new CustomOuterClassHandler(){};
+    public final Handler handler = new CustomOuterClassHandler() {
+    };
     // N.B. curly brackets '{}' are important to make the above a pseudo-anonymous class
 }
 
 class JavaInnerClassFieldReference {
     @OpenApi(responses = {@OpenApiResponse(status = "200")})
-    public final Handler handler = new CustomInnerClassHandler(ctx -> {});
+    public final Handler handler = new CustomInnerClassHandler(ctx -> {
+    });
 
     private class CustomInnerClassHandler implements Handler {
         public final int irrelevantField;
@@ -246,7 +255,9 @@ class JavaInnerClassFieldReference {
         public void handle(@NotNull Context ctx) throws Exception {
             // derived custom handler
         }
-    };
+    }
+
+    ;
 }
 
 class JavaAnonymousClassFieldReference {
@@ -309,8 +320,10 @@ public class TestOpenApiAnnotations_Java {
     }
 
     @Test
-    @Ignore // TODO: We have to either find a way to fix this, or scrap this functionality
     public void testWithJavaMethodReference() {
+        if (VersionIssuesUtil.INSTANCE.getHasIssue()) {
+            return; // TODO: We have to either find a way to fix this, or scrap this functionality
+        }
         Info info = new Info().title("Example").version("1.0.0");
         OpenApiOptions options = new OpenApiOptions(info);
 
@@ -322,8 +335,10 @@ public class TestOpenApiAnnotations_Java {
     }
 
     @Test
-    @Ignore // TODO: We have to either find a way to fix this, or scrap this functionality
     public void testWithExtendedJavaMethodReference() {
+        if (VersionIssuesUtil.INSTANCE.getHasIssue()) {
+            return; // TODO: We have to either find a way to fix this, or scrap this functionality
+        }
         Info info = new Info().title("Example").version("1.0.0");
         OpenApiOptions options = new OpenApiOptions(info);
 
@@ -335,8 +350,10 @@ public class TestOpenApiAnnotations_Java {
     }
 
     @Test
-    @Ignore // TODO: We have to either find a way to fix this, or scrap this functionality
     public void testWithJavaMethodReferenceAndMultipleMethods() {
+        if (VersionIssuesUtil.INSTANCE.getHasIssue()) {
+            return; // TODO: We have to either find a way to fix this, or scrap this functionality
+        }
         Info info = new Info().title("Example").version("1.0.0");
         OpenApiOptions options = new OpenApiOptions(info);
 
@@ -350,8 +367,10 @@ public class TestOpenApiAnnotations_Java {
     }
 
     @Test
-    @Ignore // TODO: We have to either find a way to fix this, or scrap this functionality
     public void testWithJavaMethodReferenceAndMultipleMethodsAndSamePath() {
+        if (VersionIssuesUtil.INSTANCE.getHasIssue()) {
+            return; // TODO: We have to either find a way to fix this, or scrap this functionality
+        }
         Info info = new Info().title("Example").version("1.0.0");
         OpenApiOptions options = new OpenApiOptions(info);
 
@@ -476,8 +495,10 @@ public class TestOpenApiAnnotations_Java {
     }
 
     @Test
-    @Ignore // TODO: We have to either find a way to fix this, or scrap this functionality
     public void testIfUserIsWarnedOnInvalidPath() {
+        if (VersionIssuesUtil.INSTANCE.getHasIssue()) {
+            return; // TODO: We have to either find a way to fix this, or scrap this functionality
+        }
         String log = captureStdOut(() ->
             OpenApiTestUtils.extractSchemaForTest(app -> {
                 ClassHandlerWithInvalidPath handler = new ClassHandlerWithInvalidPath();
