+++ javalin-openapi/src/test/java/io/javalin/plugin/openapi/TestWarnings.kt
@@ -1,6 +1,7 @@
 package io.javalin.plugin.openapi
 
 import io.javalin.Javalin
+import io.javalin.plugin.openapi.utils.VersionIssuesUtil
 import io.javalin.testing.TestUtil
 import io.swagger.v3.oas.models.OpenAPI
 import org.assertj.core.api.Assertions.assertThat
@@ -15,16 +16,8 @@ class TestWarnings {
                 it.registerPlugin(OpenApiPlugin(OpenApiOptions { OpenAPI() }))
             }
         }
-        val javaVersion = System.getProperty("java.version").split(".")[0].toInt()
-        val kotlinVersion = KotlinVersion.CURRENT.minor // let's face it, to JetBrains minor means major
-        val warning = when {
-            javaVersion >= 15 && kotlinVersion >= 5 -> "JDK15 and Kotlin 1.5 break reflection in different ways"
-            javaVersion >= 15 -> "JDK 15 has a breaking change to reflection"
-            kotlinVersion >= 5 -> "Kotlin 1.5 has a breaking change to reflection"
-            else -> null
-        }
-        if (warning != null) {
-            assertThat(log).contains(warning)
+        if (VersionIssuesUtil.warning != null) {
+            assertThat(log).contains(VersionIssuesUtil.warning)
         }
     }
